requirements.txt
Django==4.0.2
django-registration
djangorestframework
django-bootstrap5
django-bootstrap-modal-forms
django-impersonate
openpyxl
